Fix responsive components
