+++ src/App.js
@@ -1,8 +1,7 @@
-import { Container } from 'react-bootstrap'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
-import Header from './components/Header'
-import Footer from './components/Footer'
-import HomeScreen from './screens/HomeScreen'
+import Header from './components/Layout/Header'
+import Footer from './components/Layout/Footer'
+import Home from './pages/Home'
 
 function App() {
   return (
@@ -10,7 +9,7 @@ function App() {
       <Header />
       <main className = 'py-3'>
         <Routes>
-          <Route path='/' element={<HomeScreen/>} exact></Route>
+          <Route path='/' element={<Home />} exact></Route>
         </Routes>
       </main>
       <Footer />
