+++ src/components/Home/Carousel.js
@@ -2,7 +2,6 @@ import React from 'react';
 import {
     MDBCarousel,
     MDBCarouselItem,
-    MDBContainer,
   } from 'mdb-react-ui-kit';
 
 export default function Carousel() {
@@ -11,20 +10,20 @@ export default function Carousel() {
             <MDBCarouselItem
                 className='w-100 d-block'
                 itemId={1}
-                src={require('../images/1.jpg')}
+                src={require('../../images/1.jpg')}
                 alt='...'
                 style={{height: '60%',flex: 1,width: null}}
             />
             <MDBCarouselItem
                 className='w-100 d-block'
                 itemId={2}
-                src={require('../images/2.jpg')}
+                src={require('../../images/2.jpg')}
                 alt='...'
             />
             <MDBCarouselItem
                 className='w-100 d-block'
                 itemId={3}
-                src={require('../images/3.jpg')}
+                src={require('../../images/3.jpg')}
                 alt='...'
             />
         </MDBCarousel>
