+++ src/components/Home/TextAnimation.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Wave, Random } from 'react-animated-text';
+import { Random } from 'react-animated-text';
 
 const exampleStyle = {
-  display: 'inline-block',
+  display: 'flex',
   backgroundColor: 'black',
   color: 'white',
   border: '1px #ccc',
@@ -11,8 +11,7 @@ const exampleStyle = {
   padding: '2em 1em 1em 1em',
   width: '100%',
   fontSize: '4vh',
-  fontFamily: 'Combo', 
-  display: 'flex',
+  fontFamily: 'Combo',
   alignItems: 'centre',
   justifyContent: 'center'
 }
